100-main_opcodes

+++ 0x0F-function_pointers/100-main_opcodes.c
@@ -19,7 +19,7 @@ int main(int argc, char *argv[])
 		exit(1);
 	}
 
-	bytes = atol(argv[1]);
+	bytes = atoi(argv[1]);
 
 	if (bytes < 0)
 	{
